Create a step (central) to fetch missing dependencies
we still need to save the metadata, however we need
to download the pom to get it's sha, so we need to fuse
the flow in one: list + download => (save pom, save meta)
