+++ model/src/main/scala/ch.epfl.scala.index.model/misc/Sha1.scala
@@ -1,0 +1,8 @@
+package ch.epfl.scala.index.model.misc
+
+object Sha1 {
+  def apply(in: String): String = {
+    val md = java.security.MessageDigest.getInstance("SHA-1")
+    md.digest(in.getBytes("UTF-8")).map("%02x".format(_)).mkString
+  }
+}
